Move dispatcher.rs into lib.rs
Get rid of the dispatcher.rs file and move all of it's functions into
lib.rs, since these are what we want to expose to users.
